Move theme modifications to fork
Since we are now using a custom fork of the PaperMod theme, we also
moved the rest of the theme modifications from manage.sh to the fork.
Usage of the manage.sh script is now no longer required to launch or
deploy the blog, but we still keep it in case of other modifications
that need to be made in future.